Organized services and changed home schedule request to be based on team
